Refactor to a separate app/init.go
Resolve -> Interpolate in external packages.
